refactor targets: vgg_exporter&exporter
