+++ src/Entry/Exporter/CMakeLists.txt
@@ -3,16 +3,14 @@ include(GetGitRevisionDescription)
 if(Vulkan_FOUND)
   get_git_head_revision(REF_VAR HASH_VAR)
   message(STATUS "${REF_VAR} ${HASH_VAR}")
-  add_library(vgg_exporter src/ImageExporter.cpp)
+  add_library(vgg_exporter ImageExporter.cpp)
   target_compile_definitions(vgg_exporter PRIVATE GIT_COMMIT="${HASH_VAR}")
   set_target_properties(vgg_exporter PROPERTIES POSITION_INDEPENDENT_CODE ON)
 
   find_package(Vulkan REQUIRED)
   target_link_libraries(vgg_exporter PUBLIC vgg_layer Vulkan::Vulkan vgg_domain)
 
-
-  target_include_directories(vgg_exporter PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include)
-  target_include_directories(vgg_exporter PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/interface)
+  target_include_directories(vgg_exporter PUBLIC ${CMAKE_SOURCE_DIR}/include/VGG)
 
   install(TARGETS vgg_exporter DESTINATION lib COMPONENT vgg_module_exporter)
   install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/interface/ImageExporter.hpp DESTINATION include COMPONENT vgg_module_exporter)
